Update 16-file-formats.md
In this tutorial I was trying to find information about the $namespace and $schema sections of the documents, and that ended up distracting me from the simple message that you can add file metadata such as a format field to be more explicit about what your expecting as an input or output.

Another thing here that I was curious about is if CWLtool is actually validating the formats from the input yaml file? I think I tested it and it didn't matter mattered, so if a user can turn that on might be useful to point to the relevant documentation.