explanations about echo and file contents
* Update 02-1st-example.md

I added extra explanations about echo and the contents of the files the tutorial user should make, and the command they will execute. 

I simplified the text about 'use a YAML or JSON object in a separate file' as it is very opaque to a beginner and didn't seem immediately relevant. 

I included an explanation of how the specific command used here (cwl-runner 1st-tool.cwl echo-job.yml) is an example of the general form cwl-runner [tool-or-workflow-description] [input-job-settings].