Metadata and Authorship: Harmonize examples and highlight difference
The examples in `metadata_example2.cwl` and `metadata_example3.cwl` are
meant to be compared, and the latter is an extension of the former.
However, the differences are not obvious in the wall of text.

For this, line-highlighting is used in the latter to point out the
different parts.

In addition, minor/trivial differences in the pair are removed, so they
differ *only* in the highlighted lines. This also fixes some minor,
inconsistent indentations.