Operations: Don't imply an operation is straightforwardly useable
The doc says "[An operation] can be used like use a command line tool or
an expression", which may be read as "you can use an (abstract)
operation just like how you use a (concrete) command line tool or
expression". This is of course false and directly contradicted by the
next example.

To avoid this implication, reword it as "it can appear...".

Also, a minor clarification about "Sphinx Graphviz", which referred to a
specific processing directive in the Sphinx doc.