Fix empty section navigation sidebar and reorganize templates a bit (#471)
* fix section navigation sidebar

The sidebar generation logic in the PyData Sphinx Theme was altered, causing the navigation sidebar to become consistently displayed but always empty.

The `sidebar-nav-bs.html` override was initially added to address a responsive design issue, which has now been effectively resolved by the upstream. Let's remove the override to avoid future breakage.

`sidebar_includehidden` is set to `False` to get rid of the sidebar when its content is empty.

* reposition the project links

Add project-related links in a way that does not override upstream templates.